organized by folders, installed redux packages
