+++ khompodoevaSaiaana/components/pokemon.js
