+++ khompodoevaSaiaana/package.json
@@ -15,7 +15,9 @@
     "expo-status-bar": "~1.4.4",
     "react": "18.2.0",
     "react-native": "0.71.3",
-    "react-native-safe-area-context": "^4.5.0"
+    "react-native-safe-area-context": "^4.5.0",
+    "react-redux": "^8.0.5",
+    "redux": "^4.2.1"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
